+++ components/bridge/rpc/api/src/main/java/com/flipperdevices/bridge/rpc/api/FlipperStorageApi.kt
@@ -9,7 +9,17 @@ interface FlipperStorageApi {
 
     suspend fun delete(path: String, recursive: Boolean = false)
 
-    suspend fun upload(pathOnFlipper: String, fileOnAndroid: File, progressListener: ProgressListener)
+    suspend fun download(
+        pathOnFlipper: String,
+        fileOnAndroid: File,
+        progressListener: ProgressListener
+    )
+
+    suspend fun upload(
+        pathOnFlipper: String,
+        fileOnAndroid: File,
+        progressListener: ProgressListener
+    )
 
     suspend fun listingDirectory(pathOnFlipper: String): List<String>